Move slurm to legacy_slurm, and move grouped_slurm to slurm
