--- fractal_server/app/runner/_grouped_slurm/__init__.py
@@ -1,120 +1,0 @@
-# Copyright 2022 (C) Friedrich Miescher Institute for Biomedical Research and
-# University of Zurich
-#
-# Original authors:
-# Jacopo Nespolo <jacopo.nespolo@exact-lab.it>
-# Tommaso Comparin <tommaso.comparin@exact-lab.it>
-#
-# This file is part of Fractal and was originally developed by eXact lab S.r.l.
-# <exact-lab.it> under contract with Liberali Lab from the Friedrich Miescher
-# Institute for Biomedical Research and Pelkmans Lab from the University of
-# Zurich.
-"""
-Slurm Bakend
-
-This backend runs fractal workflows in a SLURM cluster using Clusterfutures
-Executor objects.
-"""
-from pathlib import Path
-from typing import Any
-from typing import Dict
-from typing import List
-from typing import Optional
-from typing import Union
-
-from ...models import Workflow
-from .._common import recursive_task_submission
-from ..common import async_wrap
-from ..common import TaskParameters
-from ._submit_setup import _grouped_slurm_submit_setup
-from .executor import FractalSlurmExecutor
-
-
-def _process_workflow(
-    *,
-    workflow: Workflow,
-    input_paths: List[Path],
-    output_path: Path,
-    input_metadata: Dict[str, Any],
-    logger_name: str,
-    workflow_dir: Path,
-    workflow_dir_user: Path,
-    slurm_user: Optional[str] = None,
-    worker_init: Optional[Union[str, List[str]]] = None,
-) -> Dict[str, Any]:
-    """
-    Internal processing routine for the SLURM backend
-
-    This function initialises the a FractalSlurmExecutor, setting logging,
-    workflow working dir and user to impersonate. It then schedules the
-    workflow tasks and returns the output dataset metadata.
-
-    Cf. [process_workflow][fractal_server.app.runner._local.process_workflow]
-
-    Returns:
-        output_dataset_metadata: Metadata of the output dataset
-    """
-
-    if not slurm_user:
-        raise RuntimeError(
-            "slurm_user argument is required, for slurm backend"
-        )
-
-    if isinstance(worker_init, str):
-        worker_init = worker_init.split("\n")
-
-    with FractalSlurmExecutor(
-        debug=True,
-        keep_logs=True,
-        slurm_user=slurm_user,
-        working_dir=workflow_dir,
-        working_dir_user=workflow_dir_user,
-        common_script_lines=worker_init,
-    ) as executor:
-        output_task_pars_fut = recursive_task_submission(
-            executor=executor,
-            task_list=workflow.task_list,
-            task_pars=TaskParameters(
-                input_paths=input_paths,
-                output_path=output_path,
-                metadata=input_metadata,
-            ),
-            workflow_dir=workflow_dir,
-            workflow_dir_user=workflow_dir_user,
-            submit_setup_call=_grouped_slurm_submit_setup,
-            logger_name=logger_name,
-        )
-    output_task_pars = output_task_pars_fut.result()
-    output_dataset_metadata = output_task_pars.metadata
-    return output_dataset_metadata
-
-
-async def process_workflow(
-    *,
-    workflow: Workflow,
-    input_paths: List[Path],
-    output_path: Path,
-    input_metadata: Dict[str, Any],
-    logger_name: str,
-    workflow_dir: Path,
-    workflow_dir_user: Optional[Path] = None,
-    slurm_user: Optional[str] = None,
-    worker_init: Optional[str] = None,
-) -> Dict[str, Any]:
-    """
-    Process workflow (SLURM backend public interface)
-
-    Cf. [process_workflow][fractal_server.app.runner._local.process_workflow]
-    """
-    output_dataset_metadata = await async_wrap(_process_workflow)(
-        workflow=workflow,
-        input_paths=input_paths,
-        output_path=output_path,
-        input_metadata=input_metadata,
-        logger_name=logger_name,
-        workflow_dir=workflow_dir,
-        workflow_dir_user=workflow_dir_user,
-        slurm_user=slurm_user,
-        worker_init=worker_init,
-    )
-    return output_dataset_metadata
